+++ server/component/moduleSurveyJS/ModuleSurveyJSView.php
@@ -172,7 +172,8 @@ class ModuleSurveyJSView extends BaseView
                 __DIR__ . "/js/5_survey-creator-knockout.min.js",
                 __DIR__ . "/js/6_survey-creator-core.i18n.min.js",
                 __DIR__ . "/js/7_survey.i18n.min.js",
-                __DIR__ . "/js/7_surveyjs-widgets.min.js",                
+                __DIR__ . "/js/7_surveyjs-widgets.min.js",
+                __DIR__ . "/../style/surveyJS/js/5_videoSegmentWidget.js", // Custom videoSegment question type (v1.4.7)
                 // __DIR__ . "/js/9_quill_integration.js",
                 __DIR__ . "/js/8_survey.js",                
             );
@@ -197,7 +198,8 @@ class ModuleSurveyJSView extends BaseView
                     __DIR__ . "/../style/surveyJS/css/defaultV2.min.css",
                     __DIR__ . "/css/survey-creator-core.min.css",
                     __DIR__ . "/css/quill.snow.min.css",
-                    __DIR__ . "/css/survey.css"
+                    __DIR__ . "/css/survey.css",
+                    __DIR__ . "/../style/surveyJS/css/video-segment.css" // Custom videoSegment styles (v1.4.7)
                 );
             } else {
                 $local = array(__DIR__ . "/../../../css/ext/survey-js.min.css?v=" . rtrim(shell_exec("git describe --tags")));